requirements.txt
streamlit
numpy
Pillow
streamlit-drawable-canvas
tensorflow==2.17.0
